refactor : Activity와 Fragment 종속관계에 맞춰서 파일 위치 변경
