+++ app/src/main/java/woowacourse/movie/ui/main/cinemaBottomSheet/CinemaListContract.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.ui.fragment.cinemaBottomSheet
+package woowacourse.movie.ui.main.cinemaBottomSheet
 
 import woowacourse.movie.model.CinemaState
