+++ app/src/main/java/woowacourse/movie/ui/main/reservationList/ReservationListFragment.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.ui.fragment.reservationList
+package woowacourse.movie.ui.main.reservationList
 
 import android.content.Context
 import android.os.Bundle
@@ -12,7 +12,9 @@ import woowacourse.movie.ui.adapter.ReservationListAdapter
 import woowacourse.movie.ui.confirm.ReservationConfirmActivity
 import woowacourse.movie.ui.itemModel.TicketsItemModel
 
-class ReservationListFragment : Fragment(R.layout.fragment_reservation_list), ReservationListContract.View {
+class ReservationListFragment :
+    Fragment(R.layout.fragment_reservation_list),
+    ReservationListContract.View {
     private lateinit var presenter: ReservationListContract.Presenter
 
     private var reservationRecyclerView: RecyclerView? = null
