+++ app/src/main/java/woowacourse/movie/ui/main/setting/SettingFragment.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.ui.fragment.setting
+package woowacourse.movie.ui.main.setting
 
 import android.os.Bundle
 import android.view.View
