+++ app/src/main/res/layout/fragment_movie_list.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.fragment.movieList.MovieListFragment">
+    tools:context=".ui.main.movieList.MovieListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rv_main"
