+++ app/src/test/java/woowacourse/movie/ui/main/setting/SettingPresenterTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.ui.fragment.setting
+package woowacourse.movie.ui.main.setting
 
 import io.mockk.every
 import io.mockk.mockk
